+++ src/components/popover/index.js
@@ -8,6 +8,8 @@ import Emoji from 'Components/images/emoji';
 
 import PopoverContainer from './container';
 import PopoverDialog from './dialog';
+import PopoverSearch from './search';
+import { PopoverSection, PopoverActions, PopoverInfo, PopoverTitle, InfoDescription, ActionDescription } from './base';
 import styles from './styles.styl';
 import globalStyles from '../global.styl';
 
@@ -19,27 +21,17 @@ popover pans, which have straight-walled sides rather than angled.
 ...also it's a [Bootstrap UI pattern](https://www.w3schools.com/bootstrap/bootstrap_popover.asp)
 */
 
-export { PopoverContainer, PopoverDialog };
-
-const sectionTypes = ['primary', 'secondary', 'dangerZone'];
-export const PopoverSection = ({ className, children, type }) => (
-  <section className={classnames(styles.popoverSection, styles[type], className)}>{children}</section>
-);
-
-PopoverSection.propTypes = {
-  type: PropTypes.oneOf(sectionTypes),
-  children: PropTypes.node.isRequired,
+export {
+  PopoverContainer,
+  PopoverDialog,
+  PopoverSearch,
+  PopoverSection,
+  PopoverActions,
+  PopoverInfo,
+  PopoverTitle,
+  InfoDescription,
+  ActionDescription,
 };
-
-PopoverSection.defaultProps = {
-  type: 'primary',
-};
-
-export const PopoverActions = ({ ...props }) => <PopoverSection {...props} className={styles.popoverActions} />;
-export const PopoverInfo = ({ ...props }) => <PopoverSection type="secondary" {...props} className={styles.popoverInfo} />;
-export const PopoverTitle = ({ ...props }) => <PopoverSection type="secondary" {...props} className={styles.popoverTitle} />;
-export const InfoDescription = ({ children }) => <p className={styles.infoDescription}>{children}</p>;
-export const ActionDescription = ({ children }) => <p className={styles.actionDescription}>{children}</p>;
 
 const MultiPopoverContext = createContext();
